refactor: HighlightRange 인터페이스를 Highlight/index로 이동
