+++ frontend/src/components/DatePicker/Table/Highlight/index.ts
@@ -5,6 +5,11 @@ export * from './HighlightGap';
 
 export type HighlightState = 'none' | 'startOfRange' | 'inRange' | 'endOfRange';
 
+export interface HighlightRange {
+  start: Date | null;
+  end: Date | null;
+}
+
 export interface HighlightProps extends PropsWithChildren {
   highlightState: HighlightState;
 }
