+++ frontend/src/hooks/useDatePicker/useDatePickerRange.ts
@@ -1,5 +1,5 @@
 
-import type { HighlightRange } from '@/components/DatePicker';
+import type { HighlightRange } from '@/components/DatePicker/Table/Highlight';
 import { isSameDate } from '@/utils/date';
 
 interface UseDatePickerRangeProps {
